refactor: extract keydown observer
